Issue #3193195 by abrar_arshad, fgm: avoid log failures on closures with PHP 5.x
- Also apply current (2021-01) Drupal coding standards.

+++ modules/mongodb_watchdog/src/Controller/ControllerBase.php
@@ -1,6 +1,6 @@
 <?php
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace Drupal\mongodb_watchdog\Controller;
 
@@ -157,7 +157,7 @@ abstract class ControllerBase extends CoreControllerBase {
    * @return int
    *   The actual index of the page to display.
    */
-  public static function getPage(float $count, int $requestedPage, int $height): int {
+  public static function getPage(int $count, int $requestedPage, int $height): int {
     if ($requestedPage <= 0) {
       return 0;
     }
